requirements.txt
pika
PyYAML
